todos os controlers seguros, apenas com permissao

+++ app/Http/Controllers/Admin/NotificationController.php
@@ -21,6 +21,7 @@ class NotificationController extends Controller
      */
     public function index(Request $request)
     {
+        if(!Auth::user()->isAdmin())return redirect('/dashboard')->with('flash_message', 'Acesso Não autorizado!');
         $keyword = $request->get('search');
         $perPage = 25;
         $user = Auth::user();
@@ -48,6 +49,7 @@ class NotificationController extends Controller
      */
     public function create()
     {
+        if(!Auth::user()->isAdmin())return redirect('/dashboard')->with('flash_message', 'Acesso Não autorizado!');
              return view('admin.notification.create');
     }
     
@@ -60,7 +62,7 @@ class NotificationController extends Controller
      */
     public function store(Request $request)
     {
-        
+        if(!Auth::user()->isAdmin())return redirect('/dashboard')->with('flash_message', 'Acesso Não autorizado!');
         $requestData = $request->all();
         switch($requestData['user_id']){
             case 1:
@@ -97,6 +99,7 @@ class NotificationController extends Controller
      */
     public function show($id)
     {
+        if(!Auth::user()->isAdmin())return redirect('/dashboard')->with('flash_message', 'Acesso Não autorizado!');
         $notification = Notification::findOrFail($id);
 
         return view('admin.notification.show', compact('notification'));
@@ -111,6 +114,7 @@ class NotificationController extends Controller
      */
     public function edit($id)
     {
+        if(!Auth::user()->isAdmin())return redirect('/dashboard')->with('flash_message', 'Acesso Não autorizado!');
         $notification = Notification::findOrFail($id);
         $notification->readed = 1;
         $notification->save();
@@ -128,7 +132,7 @@ class NotificationController extends Controller
      */
     public function update(Request $request, $id)
     {
-        
+        if(!Auth::user()->isAdmin())return redirect('/dashboard')->with('flash_message', 'Acesso Não autorizado!');
         $requestData = $request->all();
         
         $notification = Notification::findOrFail($id);
@@ -153,6 +157,7 @@ class NotificationController extends Controller
      */
     public function destroy($id)
     {
+        if(!Auth::user()->isAdmin())return redirect('/dashboard')->with('flash_message', 'Acesso Não autorizado!');
         Notification::destroy($id);
 
         return redirect('/admin/notification')->with('flash_message', 'Notification deleted!');
